Enable swiftcast
Able to cast a spell by pressing a button even when in weapon mode.

This works by having a flag mSwiftCast in Player and CharacterController
that will cause "spellcast" to be injected as the current weapon. When
the swiftcast is finished, the current weapon will be reverted.

I wasn't able to make OpenCS configure the string for swift casting
sSwiftCast as a game setting when editing the main Morrowind file.
However, it does show when creating a new game. As a workaround,
swiftcasting will appear as "Swift Swim" when editing the controls.
The default key is Z.