Hidden traps
Traps on doors and containers are hidden based on a dice roll.
At security skill 0 all traps are hidden, at 100 all traps are shown.

The roll is done only once per game session and applies for all doors
and containers (although those two have two separate rolls). This is to
preserve the `const' requirements for getTooltipInfo().